PIG-413: problem with float sum
git-svn-id: https://svn.apache.org/repos/asf/incubator/pig/branches/types@692279 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/builtin/FloatSum.java
@@ -83,7 +83,7 @@ public class FloatSum extends EvalFunc<Double> implements Algebraic {
                     return null;
                 }
 
-                long sum = 0;
+                double sum = 0;
                 boolean sawNonNull = false;
                 for (Iterator<Tuple> it = values.iterator(); it.hasNext();) {
                     Tuple t = (Tuple) it.next();
